Add a label field, optionally populated by (pre-bundle) token.
Originally the thought had been to have a calculated label, but it makes
more sense to have something calculated once, and a field that can be
used to filter/search in views and search_api.

Default token for address bundle is included, and in the update.

An area #1 bundle might want a manually written label?