Update geo update test module order and entity browser config
Make sure entitiy browser is installed, and set update test module as last
(does that have an effect?)

Commented out the widget config in the entity browser config in the update test
as that was causing PHP error for missing configDependencyKey. This is a known
bug with entity browser
  https://www.drupal.org/project/entity_browser/issues/2845037
and for our purposes here, we just want to check the permission is updated
to the geo entity browser.